Move update_all_linked_status_records to utils
This function works regardless of the type of operation (Lifesciences pipeline or Cloud Run job execution), and this gets rid of potential circular import / local import complications.
